+++ priv/test_repo/migrations/20260911135608_upgrade_core_v6.exs
@@ -1,0 +1,6 @@
+defmodule AuroraMeter.TestRepo.Migrations.UpgradeCoreV6 do
+  use Ecto.Migration
+
+  def up, do: AuroraMeter.Migration.up(from: 6, version: 6)
+  def down, do: AuroraMeter.Migration.down(version: 6, to: 6)
+end